+++ src/ts/checks/general/index.ts
@@ -1,4 +1,5 @@
 export { default as aprilFools } from './aprilFools';
+export { default as zip } from './zip';
 export { default as audio } from './audio';
 export { default as coverImage } from './coverImage';
 export { default as previewTime } from './previewTime';